+++ app/static/js/map.js
@@ -87,7 +87,7 @@ var clusterInfo = L.control({
 });
 
 clusterInfo.onAdd = function(map) {
-  this._div = L.DomUtil.create('div', 'info'); // create a div with a class "info"
+  this._div = L.DomUtil.create('div', 'cluster__info'); // create a div with a class "info"
   this.update();
   L.DomEvent.disableClickPropagation(this._div);
   return this._div;
@@ -102,6 +102,60 @@ clusterInfo.update = function(props) {
     '</table>';
   this._div.innerHTML
 };
+
+// this function updates the content of the clusterInfo in a centralized way
+function update_clusterInfo(properties, selectedClustersNum, clusterNum="?") {
+    console.log(properties)
+//
+    var control_content = '\
+      <div class="grid-x ">\
+          <h4 title="this is not the final style" class="cell"> Browse the settlements</h4>\
+          <div id="download_clusters" class="cell  grid-x consecutive__btn">\
+            <button class="cell large-3" style="float:left" onclick="prev_selection_fun()"> < </button> \
+            <h5 class="cell large-6">\
+                <span>' + clusterNum + ' </span> / <span id="filtered-clusters-num">\
+                ' + selectedClustersNum + '</span> \
+            </h5>\
+            <button class="cell large-3" style="float:right" onclick="next_selection_fun()"> > </button>\
+          </div>\
+      </div>';
+
+    if(properties.cluster_all_id !== undefined){
+        // all
+        control_content = control_content +
+          '<table>\
+            <tr><td align="right"><b>ID</b>:</td><td>' + properties.cluster_all_id + '</td></tr>\
+            <tr><td align="right"><b>Area</b>:</td><td>' + properties.area_km2 + '</td></tr>\
+            <tr><td align="right"><b>Distance to Grid</b>:</td><td>' + parseFloat(properties.grid_dist_km).toFixed(2) + ' km2</td></tr>\
+          </table>';
+
+          randomClusterInfo.remove();
+    };
+
+    if (properties.cluster_offgrid_id !== undefined) {
+    // og
+        control_content = control_content +
+          '<table>\
+            <tr><td align="right"><b>Area</b>:</td><td>' + parseFloat(properties.area_km2).toFixed(2) + ' km2</td></tr>\
+            <tr><td align="right"><b>Building Count</b>:</td><td>' + parseFloat(properties.building_count).toFixed(0) + '</td></tr>\
+            <tr><td align="right"><b>Building Area in km²</b>:</td><td>' + parseFloat(properties.building_area_km2).toFixed(3) + '</td></tr>\
+            <tr><td align="right"><b>Buildings per km²</b>:</td><td>' + parseFloat(properties.building_count_density_perkm2).toFixed(0) + '</td></tr>\
+            <tr><td align="right"><b>Percentage Building Area</b>:</td><td>' + parseFloat(properties.percentage_building_area).toFixed(2) + '</td></tr>\
+            <tr><td align="right"><b>Distance to Grid in km</b>:</td><td>' + parseFloat(properties.grid_dist_km).toFixed(1) + '</td></tr>\
+          </table>';
+
+          randomClusterInfo.remove();
+    };
+
+    clusterInfo.remove();
+    clusterInfo.update = function() {
+        this._div.innerHTML = control_content;
+        this._div.innerHTML;
+    };
+    // the addTo function will trigger the update() function
+    clusterInfo.addTo(map);
+};
+
 
 var randomClusterInfo = L.control({
   position: 'bottomleft'